Replaced cookies with localStorage
Lol! wrong commit message in previous commit with completely opposite meaning.

Setting SameSite was not working in chrome. Cookie was not even being saved if I will use SameSite.
And in firefox it was working but giving warning that I should its value.
Using localStorage was very easy and more efficient.